supporting overriding validate_ends_with_eos in tokenizer.call (#128)
The following changes were made to allow this:
1. In the init: change validate_ends_with_eos from str to bool, and add
another argument eos, with default="<EOS>". default behavior is
maintained. I searched for all call for the init - None used
validate_ends_with_eos
2. add a new boolean argument validate_ends_with_eos to tokenizer.call
with default None. If not None - will override
self._validate_ends_with_eos during the call.

The change was made to allow not adding <EOS> to the decoder, to allow
it to be the same length as the label. This allows cropping redundant
padding in a batch - also in decoder input and label fields